// src/app/features/cart/interfaces/ICartResponse.ts
export interface Subcategory {
  _id: string;
  name: string;
  slug: string;
  category: string;
}

export interface Category {
  _id: string;
  name: string;
  slug: string;
  category: string;
}

export interface Brand {
  _id: string;
  name: string;
  slug: string;
  category: string;
}

export interface Product {
  Subcategory: Subcategory[];
  _id: string;
  title: string;
  quantity: number;
  imageCover: string;
  Category: Category[];
  Brand: Brand[];
  ratingsAverage: number;
  id: string;
}

export interface Product {
  count: number;
  _id: string;
  product: Product;
  price: number;
}

export interface CartDetails {
  _id: string;
  cartOwner: string;
  products: Product[];
  createdAt: string;
  updatedAt: string;
  __v: number;
  totalCartPrice: number;
}

export interface ICartResponse {
  status: string;
  numOfCartItems: number;
  cartId: string;
  data: CartDetails;
}